Enable 'go to last split' in vim
The implementation is not bulletproof: for example, vim does not detect if
the mouse is used to move the focus from a vim pane to a tmux pane, and
so s:last_split_is_tmux stays False.

A possible solution would be to use a tmux hook to let vim know that its
pane regained focus, but this feels overcomplicated for now.

The plugin also won't work as expected after the creation of a vim
split; if the 'last split' prior to the creation of the new split was a
tmux pane, this pane will remain the 'last split' after window creation.